Moves the MixHelper module to the support folder
